+++ server/src/completion.ts
@@ -27,7 +27,7 @@ export async function getCompletionItems(
 			return null
 		}
 
-		return await getCompletionItemsInHTML(fromPart, currentHTMLService, document, cssServiceMap, offset)
+		return await getCompletionItemsInHTML(fromPart, currentHTMLService, document, cssServiceMap, configuration, offset)
 	}
 	else if (isCSSFile) {
 		let currentCSSService = await cssServiceMap.forceGetServiceByDocument(document)
@@ -53,6 +53,7 @@ async function getCompletionItemsInHTML(
 	currentService: HTMLService,
 	document: TextDocument,
 	cssServiceMap: CSSServiceMap,
+	configuration: Configuration,
 	offset: number
 ): Promise<CompletionItem[] | null> {
 
@@ -63,8 +64,8 @@ async function getCompletionItemsInHTML(
 
 	// Complete html element class name.
 	if (fromPart.isHTMLType()) {
-		labels.add(CompletionLabelType.Definition, currentService.getCompletionLabels(matchPart, fromPart))
-		labels.add(CompletionLabelType.Definition, await cssServiceMap.getCompletionLabels(matchPart, fromPart))
+		labels.add(CompletionLabelType.Definition, currentService.getCompletionLabels(matchPart, fromPart, configuration.maxHoverStylePropertyCount))
+		labels.add(CompletionLabelType.Definition, await cssServiceMap.getCompletionLabels(matchPart, fromPart, configuration.maxHoverStylePropertyCount))
 	}
 
 	// Complete class name for css selector of a css document.
@@ -74,7 +75,7 @@ async function getCompletionItemsInHTML(
 
 		// Find all css variable declarations across all css documents.
 		if (fromPart.isCSSVariableType()) {
-			labels.add(CompletionLabelType.CSSVariable, await cssServiceMap.getCompletionLabels(matchPart, fromPart))
+			labels.add(CompletionLabelType.CSSVariable, await cssServiceMap.getCompletionLabels(matchPart, fromPart, configuration.maxHoverStylePropertyCount))
 		}
 	}
 
@@ -95,7 +96,6 @@ async function getCompletionItemsInCSS(
 	let matchPart = PartConvertor.toDefinitionMode(fromPart)
 	let labels = new CompletionLabels()
 
-
 	// Find selector referenced completions from current document, and across all html documents.
 	// It ignores css selectors declaration completions, which will be filled by vscode.
 	if (fromPart.isSelectorType()) {
@@ -104,7 +104,9 @@ async function getCompletionItemsInCSS(
 
 	// Find all css variable declarations across all css documents.
 	else if (fromPart.isCSSVariableType()) {
-		labels.add(CompletionLabelType.CSSVariable, await cssServiceMap.getCompletionLabels(matchPart, fromPart))
+		labels.add(CompletionLabelType.CSSVariable,
+			await cssServiceMap.getCompletionLabels(matchPart, fromPart, configuration.maxHoverStylePropertyCount)
+		)
 
 		// Remove repetitive items with current document.
 		if (configuration.disableOwnCSSVariableCompletion) {