+++ server/src/languages/services/types.ts
@@ -1,0 +1,4 @@
+export interface CompletionLabel {
+	text: string | undefined
+	markdown: string | undefined
+}